+++ go-plugin/mcpassert.go
@@ -191,9 +191,3 @@ func mergeOpts(opts []Options) Options {
 	return opt
 }
 
-func min(a, b int) int {
-	if a < b {
-		return a
-	}
-	return b
-}